+++ application/api/handler/alarm.go
@@ -1,19 +1,19 @@
-package handler2
+package handler
 
 import (
 	"github.com/labstack/echo/v4"
 	"github.com/takoikatakotako/charalarm-api/entity/request"
 	"github.com/takoikatakotako/charalarm-api/entity/response"
-	"github.com/takoikatakotako/charalarm-api/service2"
+	"github.com/takoikatakotako/charalarm-api/service"
 	"github.com/takoikatakotako/charalarm-api/util/auth"
 	"net/http"
 )
 
-type User struct {
-	Service service2.User
+type Alarm struct {
+	Service service.Alarm
 }
 
-func (u *User) UserInfoGet(c echo.Context) error {
+func (a *Alarm) AlarmListGet(c echo.Context) error {
 	authorizationHeader := c.Request().Header.Get("Authorization")
 	userID, authToken, err := auth.Basic(authorizationHeader)
 	if err != nil {
@@ -21,31 +21,16 @@ func (u *User) UserInfoGet(c echo.Context) error {
 		return c.JSON(http.StatusInternalServerError, res)
 	}
 
-	res, err := u.Service.GetUser(userID, authToken)
+	res, err := a.Service.GetAlarmList(userID, authToken)
 	if err != nil {
 		res := response.Message{Message: "Error!"}
 		return c.JSON(http.StatusInternalServerError, res)
 	}
+
 	return c.JSON(http.StatusOK, res)
 }
 
-func (u *User) UserSignupPost(c echo.Context) error {
-	req := new(request.UserSignUp)
-	if err := c.Bind(&req); err != nil {
-		res := response.Message{Message: "Error!"}
-		return c.JSON(http.StatusInternalServerError, res)
-	}
-	ipAddress := c.RealIP()
-
-	res, err := u.Service.Signup(req.UserID, req.AuthToken, req.Platform, ipAddress)
-	if err != nil {
-		res := response.Message{Message: "Error!"}
-		return c.JSON(http.StatusInternalServerError, res)
-	}
-	return c.JSON(http.StatusOK, res)
-}
-
-func (u *User) UserUpdatePremiumPost(c echo.Context) error {
+func (a *Alarm) AlarmAddPost(c echo.Context) error {
 	authorizationHeader := c.Request().Header.Get("Authorization")
 	userID, authToken, err := auth.Basic(authorizationHeader)
 	if err != nil {
@@ -53,25 +38,24 @@ func (u *User) UserUpdatePremiumPost(c echo.Context) error {
 		return c.JSON(http.StatusInternalServerError, res)
 	}
 
-	req := new(request.UserUpdatePremiumPlan)
+	req := new(request.AddAlarmRequest)
 	if err := c.Bind(&req); err != nil {
 		res := response.Message{Message: "Error!"}
 		return c.JSON(http.StatusInternalServerError, res)
 	}
 
-	err = u.Service.UpdatePremiumPlan(userID, authToken, req.EnablePremiumPlan)
+	err = a.Service.AddAlarm(userID, authToken, req.Alarm)
 	if err != nil {
 		res := response.Message{Message: "Error!"}
 		return c.JSON(http.StatusInternalServerError, res)
 	}
-
 	res := response.Message{
-		Message: Healthy,
+		Message: "Health",
 	}
 	return c.JSON(http.StatusOK, res)
 }
 
-func (u *User) UserWithdrawPost(c echo.Context) error {
+func (a *Alarm) AlarmEditPost(c echo.Context) error {
 	authorizationHeader := c.Request().Header.Get("Authorization")
 	userID, authToken, err := auth.Basic(authorizationHeader)
 	if err != nil {
@@ -79,14 +63,44 @@ func (u *User) UserWithdrawPost(c echo.Context) error {
 		return c.JSON(http.StatusInternalServerError, res)
 	}
 
-	err = u.Service.Withdraw(userID, authToken)
+	req := new(request.AddAlarmRequest)
+	if err := c.Bind(&req); err != nil {
+		res := response.Message{Message: "Error!"}
+		return c.JSON(http.StatusInternalServerError, res)
+	}
+
+	err = a.Service.EditAlarm(userID, authToken, req.Alarm)
+	if err != nil {
+		res := response.Message{Message: "Error!"}
+		return c.JSON(http.StatusInternalServerError, res)
+	}
+	res := response.Message{
+		Message: "Health",
+	}
+	return c.JSON(http.StatusOK, res)
+}
+
+func (a *Alarm) AlarmDeletePost(c echo.Context) error {
+	authorizationHeader := c.Request().Header.Get("Authorization")
+	userID, authToken, err := auth.Basic(authorizationHeader)
 	if err != nil {
 		res := response.Message{Message: "Error!"}
 		return c.JSON(http.StatusInternalServerError, res)
 	}
 
+	req := new(request.DeleteAlarmRequest)
+	if err := c.Bind(&req); err != nil {
+		res := response.Message{Message: "Error!"}
+		return c.JSON(http.StatusInternalServerError, res)
+	}
+
+	err = a.Service.DeleteAlarm(userID, authToken, req.AlarmID)
+	if err != nil {
+		res := response.Message{Message: "Error!"}
+		return c.JSON(http.StatusInternalServerError, res)
+	}
 	res := response.Message{
-		Message: Healthy,
+		Message: "Health",
 	}
 	return c.JSON(http.StatusOK, res)
 }
