+++ application/api/handler/healthcheck.go
@@ -1,4 +1,4 @@
-package handler2
+package handler
 
 import (
 	"github.com/labstack/echo/v4"
