+++ application/api/handler/user.go
@@ -1,19 +1,19 @@
-package handler2
+package handler
 
 import (
 	"github.com/labstack/echo/v4"
 	"github.com/takoikatakotako/charalarm-api/entity/request"
 	"github.com/takoikatakotako/charalarm-api/entity/response"
-	"github.com/takoikatakotako/charalarm-api/service2"
+	"github.com/takoikatakotako/charalarm-api/service"
 	"github.com/takoikatakotako/charalarm-api/util/auth"
 	"net/http"
 )
 
-type Alarm struct {
-	Service service2.Alarm
+type User struct {
+	Service service.User
 }
 
-func (a *Alarm) AlarmListGet(c echo.Context) error {
+func (u *User) UserInfoGet(c echo.Context) error {
 	authorizationHeader := c.Request().Header.Get("Authorization")
 	userID, authToken, err := auth.Basic(authorizationHeader)
 	if err != nil {
@@ -21,16 +21,31 @@ func (a *Alarm) AlarmListGet(c echo.Context) error {
 		return c.JSON(http.StatusInternalServerError, res)
 	}
 
-	res, err := a.Service.GetAlarmList(userID, authToken)
+	res, err := u.Service.GetUser(userID, authToken)
 	if err != nil {
 		res := response.Message{Message: "Error!"}
 		return c.JSON(http.StatusInternalServerError, res)
 	}
-
 	return c.JSON(http.StatusOK, res)
 }
 
-func (a *Alarm) AlarmAddPost(c echo.Context) error {
+func (u *User) UserSignupPost(c echo.Context) error {
+	req := new(request.UserSignUp)
+	if err := c.Bind(&req); err != nil {
+		res := response.Message{Message: "Error!"}
+		return c.JSON(http.StatusInternalServerError, res)
+	}
+	ipAddress := c.RealIP()
+
+	res, err := u.Service.Signup(req.UserID, req.AuthToken, req.Platform, ipAddress)
+	if err != nil {
+		res := response.Message{Message: "Error!"}
+		return c.JSON(http.StatusInternalServerError, res)
+	}
+	return c.JSON(http.StatusOK, res)
+}
+
+func (u *User) UserUpdatePremiumPost(c echo.Context) error {
 	authorizationHeader := c.Request().Header.Get("Authorization")
 	userID, authToken, err := auth.Basic(authorizationHeader)
 	if err != nil {
@@ -38,24 +53,25 @@ func (a *Alarm) AlarmAddPost(c echo.Context) error {
 		return c.JSON(http.StatusInternalServerError, res)
 	}
 
-	req := new(request.AddAlarmRequest)
+	req := new(request.UserUpdatePremiumPlan)
 	if err := c.Bind(&req); err != nil {
 		res := response.Message{Message: "Error!"}
 		return c.JSON(http.StatusInternalServerError, res)
 	}
 
-	err = a.Service.AddAlarm(userID, authToken, req.Alarm)
+	err = u.Service.UpdatePremiumPlan(userID, authToken, req.EnablePremiumPlan)
 	if err != nil {
 		res := response.Message{Message: "Error!"}
 		return c.JSON(http.StatusInternalServerError, res)
 	}
+
 	res := response.Message{
-		Message: "Health",
+		Message: Healthy,
 	}
 	return c.JSON(http.StatusOK, res)
 }
 
-func (a *Alarm) AlarmEditPost(c echo.Context) error {
+func (u *User) UserWithdrawPost(c echo.Context) error {
 	authorizationHeader := c.Request().Header.Get("Authorization")
 	userID, authToken, err := auth.Basic(authorizationHeader)
 	if err != nil {
@@ -63,44 +79,14 @@ func (a *Alarm) AlarmEditPost(c echo.Context) error {
 		return c.JSON(http.StatusInternalServerError, res)
 	}
 
-	req := new(request.AddAlarmRequest)
-	if err := c.Bind(&req); err != nil {
-		res := response.Message{Message: "Error!"}
-		return c.JSON(http.StatusInternalServerError, res)
-	}
-
-	err = a.Service.EditAlarm(userID, authToken, req.Alarm)
+	err = u.Service.Withdraw(userID, authToken)
 	if err != nil {
 		res := response.Message{Message: "Error!"}
 		return c.JSON(http.StatusInternalServerError, res)
 	}
+
 	res := response.Message{
-		Message: "Health",
-	}
-	return c.JSON(http.StatusOK, res)
-}
-
-func (a *Alarm) AlarmDeletePost(c echo.Context) error {
-	authorizationHeader := c.Request().Header.Get("Authorization")
-	userID, authToken, err := auth.Basic(authorizationHeader)
-	if err != nil {
-		res := response.Message{Message: "Error!"}
-		return c.JSON(http.StatusInternalServerError, res)
-	}
-
-	req := new(request.DeleteAlarmRequest)
-	if err := c.Bind(&req); err != nil {
-		res := response.Message{Message: "Error!"}
-		return c.JSON(http.StatusInternalServerError, res)
-	}
-
-	err = a.Service.DeleteAlarm(userID, authToken, req.AlarmID)
-	if err != nil {
-		res := response.Message{Message: "Error!"}
-		return c.JSON(http.StatusInternalServerError, res)
-	}
-	res := response.Message{
-		Message: "Health",
+		Message: Healthy,
 	}
 	return c.JSON(http.StatusOK, res)
 }
