+++ application/api/repository/aws_dynamodb_alarm.go
@@ -1,4 +1,4 @@
-package repository2
+package repository
 
 import (
 	"context"
