+++ application/api/repository/environment.go
@@ -1,4 +1,4 @@
-package repository2
+package repository
 
 import (
 	"errors"
