+++ application/api/service/alarm.go
@@ -1,11 +1,11 @@
-package service2
+package service
 
 import (
 	"errors"
 	"fmt"
 	"github.com/takoikatakotako/charalarm-api/entity/request"
 	"github.com/takoikatakotako/charalarm-api/entity/response"
-	"github.com/takoikatakotako/charalarm-api/repository2"
+	"github.com/takoikatakotako/charalarm-api/repository"
 	"github.com/takoikatakotako/charalarm-api/util/converter"
 	"github.com/takoikatakotako/charalarm-api/util/logger"
 	"github.com/takoikatakotako/charalarm-api/util/message"
@@ -17,7 +17,7 @@ const (
 )
 
 type Alarm struct {
-	AWS repository2.AWS
+	AWS repository.AWS
 }
 
 // AddAlarm アラームを追加
